Added support for spell checking comments and strings.
In the diff viewer, strings and comments are now checked for spelling errors.
These will appear with a red underline and, when clicked, the word can either
be added to the dictionary or ignored.

This makes it easy to detect typos before customers have a chance to see
them. We may later make it easy to just report spelling errors with a couple
clicks.

This requires the Python enchant module for the spell checking to function.