Improve spell checking reliability.
Spell checking is now turned off for some file types. Specifically,
those that are common XML/SGML-based files. Having spell checking in
attributes does very little for us in these cases.

We're also a lot smarter now about highlighting the spelling errors.
Originally, the generated HTML was being processed, and such things as
a tag or entity could be seen as a misspelled word and modified, causing
some breakages in our generated HTML.

Now we use the same function to highlight regions in HTML (for the inter-line
diffs). The original set of plain text lines gets passed in, and we do the
spell checking on these, highlighting the matching regions in the HTML.